+++ Adapter/PlentymarketsAdapter/ResponseParser/Product/ProductResponseParser.php
@@ -122,7 +122,7 @@ class ProductResponseParser implements ProductResponseParserInterface
         $productObject = new Product();
         $productObject->setIdentifier($identity->getObjectIdentifier());
         $productObject->setName((string) $product['texts'][0]['name1']);
-        $productObject->setActive($this->getActive($variations));
+        $productObject->setActive($this->getActive($variations, $mainVariation));
         $productObject->setNumber($this->getProductNumber($variations));
         $productObject->setShopIdentifiers($shopIdentifiers);
         $productObject->setManufacturerIdentifier($this->getManufacturerIdentifier($product));
@@ -675,11 +675,16 @@ class ProductResponseParser implements ProductResponseParserInterface
 
     /**
      * @param Variation[] $variations
+     * @param array $mainVariation
      *
      * @return bool
      */
-    private function getActive(array $variations = [])
+    private function getActive(array $variations = [], array $mainVariation)
     {
+        if (!$mainVariation['isActive']) {
+            return false;
+        }
+
         foreach ($variations as $variation) {
             if ($variation->getActive()) {
                 return true;